Improved: timewalking, heartstone quests, gm visibility, itemlevel check, first kill system. Implemented: Dimension System

+++ modules/PlayerStats/src/classes/AzthGroupMgr.h
@@ -10,7 +10,7 @@ public:
 
     void saveToDb();
 
-    uint8 levelMaxGroup;
+    uint32 levelMaxGroup;
     Group* group;
 };
 #endif
